+++ app/inventory/src/main/kotlin/org/ivcode/inventory/repository/model/CheckoutEntity.kt
@@ -3,8 +3,8 @@ package org.ivcode.inventory.repository.model
 import java.time.LocalDateTime
 
 data class CheckoutEntity(
-    val checkoutId: Int? = null,
-    val assetId: Int? = null,
+    val checkoutId: Long? = null,
+    val assetId: Long? = null,
     val userDisplayName: String? = null,
     val notes: String? = null,
     val timestamp: LocalDateTime? = null
