+++ app/inventory/src/main/kotlin/org/ivcode/inventory/repository/model/ImageInfoEntity.kt
@@ -1,8 +1,8 @@
 package org.ivcode.inventory.repository.model
 
 data class ImageInfoEntity (
-    val imageId: Int? = null,
-    val assetId: Int? = null,
+    val imageId: Long? = null,
+    val assetId: Long? = null,
     val filename: String? = null,
     val mime: String? = null
 )
