+++ app/inventory/src/main/kotlin/org/ivcode/inventory/service/model/AssetSummary.kt
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "An asset summary")
 data class AssetSummary (
-    val assetId: Int,
+    val assetId: Long,
     val type: AssetType,
     val name: String,
     val barcode: String?
