+++ app/inventory/src/main/kotlin/org/ivcode/inventory/service/model/ImageInfo.kt
@@ -1,8 +1,8 @@
 package org.ivcode.inventory.service.model
 
 data class ImageInfo (
-    val imageId: Int,
-    val assetId: Int,
+    val imageId: Long,
+    val assetId: Long,
     val filename: String,
     val mime: String
 )
