+++ app/inventory/src/main/kotlin/org/ivcode/inventory/service/model/NavigationElement.kt
@@ -1,7 +1,7 @@
 package org.ivcode.inventory.service.model
 
 data class NavigationElement (
-    val groupId: Int?,
+    val groupId: Long?,
     val name: String,
     val path: List<GroupPathElement>,
     val groups: List<GroupSummary>,
